Refactor. code for floating consultation button and test cases
Co-authored By: NikhithaKom-TW <nikhitha.k@thoughtworks.com>
