+++ src/app/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import ConsultationNotes from '../components/consulatation-notes/consultation-notes'
+import ConsultationNotes from '../components/consultation-notes/consultation-notes'
 
 const App = () => {
   return <ConsultationNotes />
