--- src/components/consultation-pad-container/consultation-pad-container.scss
@@ -1,25 +1,0 @@
-//  @import '@carbon/react';
-.saveButton{
-  left: 30%
-}
-
-.microphoneIcon {
-  fill: rgb(240, 18, 18);
-  height: 55px;
-  width: 50px;
-}
-
-.stopIcon {
-  fill: rgb(240, 18, 18);
-  height: 55px;
-  width: 75px;
-}
-
-.padBottomArea{
-    display: flex;
-    padding: 1rem 0rem 0rem 0rem;
-    align-items: center;
-    justify-content: center;
-  }
-
-
