+++ src/components/consultation-pad/consultation-pad.test.tsx
@@ -3,12 +3,12 @@ import React from 'react'
 import {ConsultationPad} from './consultation-pad'
 
 describe('Consultation Pad', () => {
-  it('should show Consultation Notes heading when Consultation pad component is rendered', () => {
-    render(<ConsultationPad isConsultationPadClosed={false} />)
+  it('should show consultation notes heading, close button, start mic and save button when consultation pad component is rendered', () => {
+    render(<ConsultationPad setConsultationPadToClosed={true} />)
+
     expect(screen.getByText('Consultation Notes')).toBeInTheDocument()
-  })
-  it('should show start mic when Consultation pad component is rendered', () => {
-    render(<ConsultationPad isConsultationPadClosed={false} />)
+    expect(screen.getByRole('button', {name: /close/i}))
     expect(screen.getByLabelText('Start Mic')).toBeInTheDocument()
+    expect(screen.getByRole('button', {name: /Save/i})).toBeDisabled()
   })
 })
