+++ src/components/consultation-pad/consultation-pad.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import {ConsultationPadContents} from '../consultation-pad-contents/consultation-pad-contents'
 import {Modal} from '@carbon/react'
-export const ConsultationPad = ({isConsultationPadClosed}) => {
+export const ConsultationPad = ({setConsultationPadToClosed}) => {
   return (
     <>
       <Modal
         modalHeading="Consultation Notes"
         open={true}
-        onRequestClose={() => isConsultationPadClosed(true)}
+        onRequestClose={() => setConsultationPadToClosed(false)}
         passiveModal={true}
       >
         <ConsultationPadContents />
